Extract diagnostics test and compilation infos from unit test collection
- Diagnostic information are now in diagnostics.test.cs
- Extension.test.ts is renamed to compilation_information.test.ts
